automation/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
appium-python-client==3.1.1
schedule==1.2.1

